Minor cleanup in Processes chapter
No double-list is generated. Since a sub-chapter list is automatically
created, let's keep that one, and make it nicer (i.e. strike through
old processes).

To implement the above, the SimpleText node was improved with added
style flags.

The Italic class was also added, but not used at all yet. However
as long as we have Bold and Strike classes, we should have Italic as
well. If we remove one of them, we should remove all three of them.